Fix circular dependencies on c125_annotations branch (#3636)
* Fixed circular dependencies
* moved forceselection to a utils file
* fix travis build
