+++ web/client/components/data/identify/viewers/TemplateViewer.jsx
@@ -8,7 +8,7 @@
 
 const React = require('react');
 const {template} = require('lodash');
-const MapInfoUtils = require('../../../../utils/MapInfoUtils');
+const TemplateUtils = require('../../../../utils/TemplateUtils');
 const HtmlRenderer = require('../../../misc/HtmlRenderer');
 const {Row, Col, Grid} = require('react-bootstrap');
 
@@ -17,7 +17,7 @@ module.exports = ({layer = {}, response}) => (
     {response.features.map((feature, i) =>
         <Row key={i}>
             <Col xs={12}>
-                <HtmlRenderer html={template(MapInfoUtils.getCleanTemplate(layer.featureInfo && layer.featureInfo.template || '', feature, /\$\{.*?\}/g, 2, 1))(feature)}/>
+                <HtmlRenderer html={template(TemplateUtils.getCleanTemplate(layer.featureInfo && layer.featureInfo.template || '', feature, /\$\{.*?\}/g, 2, 1))(feature)}/>
             </Col>
             <Col xs={12}>
                 <hr/>
