+++ web/client/components/map/openlayers/DrawSupport.jsx
@@ -17,7 +17,8 @@ const {isSimpleGeomType, getSimpleGeomType} = require('../../../utils/MapUtils')
 const {reprojectGeoJson, calculateDistance, reproject} = require('../../../utils/CoordinatesUtils');
 const {createStylesAsync} = require('../../../utils/VectorStyleUtils');
 const wgs84Sphere = new ol.Sphere(6378137);
-const {transformPolygonToCircle, isCompletePolygon} = require('../../../utils/DrawSupportUtils');
+const {transformPolygonToCircle} = require('../../../utils/DrawSupportUtils');
+const {isCompletePolygon} = require('../../../utils/AnnotationsUtils');
 const VectorStyle = require('./VectorStyle');
 const {parseStyles} = require('./VectorStyle');
 const geojsonFormat = new ol.format.GeoJSON();
