+++ web/client/utils/TemplateUtils.js
@@ -6,9 +6,9 @@
  * LICENSE file in the root directory of this source tree.
  */
 
-const {isString} = require('lodash');
+const {isString, has, trim} = require('lodash');
 
-module.exports = {
+const TemplateUtils = {
     /**
      * generates a template string to use for static replacements.
      * It's useful for using a similar syntax for static configured strings to
@@ -56,5 +56,33 @@ module.exports = {
                 callback(null, e);
             }
         });
+    },
+    /**
+     * check if a string attribute is inside of a given object
+     * @param feature {object}
+     * @param attribute {string} name of attribue with dot notations
+     * @param start {array} substring start
+     * @param end {array} substring end
+     * @return {bool} true if feature contains the attribute
+     */
+    validateStringAttribute: (feature, attribute, start = 0, end = 0) => {
+        const path = isString(attribute) && trim(attribute.substring(start, attribute.length - end)) || '';
+        return has(feature, path);
+    },
+    /**
+     * returns a valid template
+     * @param template {string} text with attribute to validate
+     * @param feature {object} object to match attributes
+     * @param regex {regex}
+     * @param start {array} substring start
+     * @param end {array} substring end
+     * @return {string} templete without invalid attribute and html tag inside attribute, e.g. ${ <p>properties.id</p> } -> ${ properties.id }
+     */
+    getCleanTemplate: (template, feature, regex, start = 0, end = 0) => {
+        const matchVariables = isString(template) && template.match(regex);
+        const replacedTag = matchVariables && matchVariables.map(temp => ({ previous: temp, next: TemplateUtils.validateStringAttribute(feature, temp.replace(/(<([^>]+)>)/ig, ''), start, end) && temp.replace(/(<([^>]+)>)/ig, '') || ''})) || null;
+        return replacedTag && replacedTag.reduce((temp, variable) => temp.replace(variable.previous, variable.next), template) || template || '';
     }
 };
+
+module.exports = TemplateUtils;
