+++ web/client/utils/__tests__/DrawSupportUtils-test.js
@@ -7,7 +7,7 @@
  */
 
 const expect = require('expect');
-const {boundsToOLExtent, fromLeafletFeatureToQueryform, isCompletePolygon} = require('../DrawSupportUtils');
+const {boundsToOLExtent, fromLeafletFeatureToQueryform} = require('../DrawSupportUtils');
 const L = require('leaflet');
 
 describe('LocaleUtils', () => {
@@ -42,17 +42,5 @@ describe('LocaleUtils', () => {
         const bounds = L.latLngBounds([[1, 2], [3, 4]]);
         const convertedBounds = boundsToOLExtent(bounds);
         expect(convertedBounds.length).toBe(4);
-    });
-
-    it('test isCompletePolygon defaults', () => {
-        const polygonCoords1 = [[[1, 1], [2, 2]]];
-        const polygonCoords2 = [[[1, 1], [2, 2], [1, 1]]];
-        const polygonCoords3 = [[[1, 1], [2, 2], [3, 3], [1, 1]]];
-        const polygonCoords4 = [[[1, 1], [2, undefined], [3, 3], [1, 1]]];
-        expect(isCompletePolygon()).toBe(false);
-        expect(isCompletePolygon(polygonCoords1)).toBe(false);
-        expect(isCompletePolygon(polygonCoords2)).toBe(false);
-        expect(isCompletePolygon(polygonCoords3)).toBe(true);
-        expect(isCompletePolygon(polygonCoords4)).toBe(false);
     });
 });
